be consistent in time_t handling
cast time_t values with (long long) and print them with %lld. Makes time_t work
on 64-bit time_t values and is consistent throughout the code.
Fix few actual issues like assuming time_t == long.

+++ cf-agent/verify_files_hashes.c
@@ -403,7 +403,7 @@ void LogHashChange(const char *file, FileState status, char *msg, Promise *pp)
 
     const char *handle = PromiseID(pp);
 
-    fprintf(fp, "%ld,%s,%s,%c,%s\n", (long) now, handle, file, FileStateToChar(status), msg);
+    fprintf(fp, "%lld,%s,%s,%c,%s\n", (long long) now, handle, file, FileStateToChar(status), msg);
     fclose(fp);
 
     chmod(fname, perm);
