+++ tests/unit/connection_management_test.c
@@ -18,13 +18,13 @@ static void test_purge_old_connections_nochange(void)
     Item *connections = NULL;
     char time_str[64];
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS);
     PrependItem(&connections, "123.123.123.3", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS + 1);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS + 1);
     PrependItem(&connections, "123.123.123.2", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS + 100);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS + 100);
     PrependItem(&connections, "123.123.123.1", time_str);
 
     assert_int_equal(ListLen(connections), 3);
@@ -48,13 +48,13 @@ static void test_purge_old_connections_purge_first(void)
     Item *connections = NULL;
     char time_str[64];
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS + 100);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS + 100);
     PrependItem(&connections, "123.123.123.3", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS + 2);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS + 2);
     PrependItem(&connections, "123.123.123.2", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS - 5);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS - 5);
     PrependItem(&connections, "123.123.123.1", time_str);
 
     assert_int_equal(ListLen(connections), 3);
@@ -78,13 +78,13 @@ static void test_purge_old_connections_purge_middle(void)
     Item *connections = NULL;
     char time_str[64];
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS);
     PrependItem(&connections, "123.123.123.3", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS - 1);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS - 1);
     PrependItem(&connections, "123.123.123.2", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS + 100);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS + 100);
     PrependItem(&connections, "123.123.123.1", time_str);
 
     assert_int_equal(ListLen(connections), 3);
@@ -108,13 +108,13 @@ static void test_purge_old_connections_purge_last(void)
     Item *connections = NULL;
     char time_str[64];
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS - 100);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS - 100);
     PrependItem(&connections, "123.123.123.3", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS + 10);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS + 10);
     PrependItem(&connections, "123.123.123.2", time_str);
 
-    snprintf(time_str, sizeof(time_str), "%ld", time_now - CONNECTION_MAX_AGE_SECONDS);
+    snprintf(time_str, sizeof(time_str), "%lld", (long long)time_now - CONNECTION_MAX_AGE_SECONDS);
     PrependItem(&connections, "123.123.123.1", time_str);
 
     assert_int_equal(ListLen(connections), 3);
